requirements.txt
numpy<2
opencv-python==4.13.0.92
Pillow==12.1.1
PySide6==6.7.3
shiboken6==6.7.3
pyserial==3.5
